+++ sdk/storage/Azure.Storage.Blobs/src/BlobBaseClient.cs
@@ -5068,7 +5068,7 @@ namespace Azure.Storage.Blobs.Specialized
         /// If multiple failures occur, an <see cref="AggregateException"/> will be thrown,
         /// containing each failure instance.
         /// </remarks>
-        internal async Task<Response<BlobProperties>> GetPropertiesInternal(
+        internal virtual async Task<Response<BlobProperties>> GetPropertiesInternal(
             BlobRequestConditions conditions,
             bool async,
             RequestContext context,